Add dist files for release

+++ resources/build.sh
@@ -3,6 +3,11 @@
 set -e
 set -o pipefail
 
+if [ ! -d "node_modules/.bin" ]; then
+  echo "Be sure to run \`npm install\` before building GraphiQL."
+  exit 1
+fi
+
 rm -rf dist/ && mkdir -p dist/
 babel src --ignore __tests__ --out-dir dist/
 echo "Bundling graphiql.js..."
